+++ src/main/java/com/playus/userservice/domain/notification/specification/NotificationControllerSpecification.java
@@ -2,8 +2,6 @@ package com.playus.userservice.domain.notification.specification;
 
 import com.playus.userservice.domain.notification.dto.response.NotificationResponse;
 import com.playus.userservice.domain.oauth.dto.CustomOAuth2User;
-import com.playus.userservice.domain.user.feign.response.CommentNotificationEvent;
-import com.playus.userservice.domain.user.feign.response.PartyNotificationEvent;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -21,12 +19,14 @@ import java.util.List;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.TEXT_EVENT_STREAM_VALUE;
 
-@Tag(name = "Notification", description = "실시간 알림 및 알림 관리 API")
+@Tag(name = "Notification", description = "실시간 알림 스트림 및 조회 API")
 public interface NotificationControllerSpecification {
+
+    /* =========================  SSE 구독  ========================= */
 
     @Operation(
             summary     = "SSE 구독",
-            description = "Server-Sent Events 로 실시간 알림을 수신하기 위한 스트림을 시작합니다.",
+            description = "Server-Sent Events 스트림을 시작하여 실시간 알림을 수신합니다.",
             security    = @SecurityRequirement(name = "AccessCookie"),
             parameters  = {
                     @Parameter(
@@ -38,7 +38,7 @@ public interface NotificationControllerSpecification {
                     ),
                     @Parameter(
                             name        = "Last-Event-ID",
-                            description = "클라이언트가 보관한 마지막 이벤트 ID (연결 복구용)",
+                            description = "클라이언트가 저장한 마지막 이벤트 ID (재연결 시)",
                             in          = ParameterIn.HEADER,
                             required    = false,
                             example     = "1_1715151515151"
@@ -46,36 +46,29 @@ public interface NotificationControllerSpecification {
             }
     )
     @ApiResponses({
-            @ApiResponse(responseCode = "200",
-                    description  = "연결 성공 (Event Stream 반환)",
-                    content      = @Content(mediaType = TEXT_EVENT_STREAM_VALUE)),
-            @ApiResponse(responseCode = "401",
-                    description  = "인증 실패",
-                    content      = @Content(
+            @ApiResponse(
+                    responseCode = "200", description = "연결 성공",
+                    content = @Content(mediaType = TEXT_EVENT_STREAM_VALUE)
+            ),
+            @ApiResponse(
+                    responseCode = "401", description = "인증 실패",
+                    content = @Content(
                             mediaType = APPLICATION_JSON_VALUE,
                             examples  = @ExampleObject(value = """
                             {
-                                "code": 401,
-                                "status": "UNAUTHORIZED",
-                                "message": "유효하지 않은 토큰입니다."
+                              "code": 401,
+                              "status": "UNAUTHORIZED",
+                              "message": "유효하지 않은 토큰입니다."
                             }""")
-                    )),
-            @ApiResponse(responseCode = "500",
-                    description  = "서버 내부 오류",
-                    content      = @Content(
-                            mediaType = APPLICATION_JSON_VALUE,
-                            examples  = @ExampleObject(value = """
-                            {
-                                "code": 500,
-                                "status": "INTERNAL_SERVER_ERROR",
-                                "message": "서버 내부 오류가 발생했습니다. 관리자에게 문의해 주세요."
-                            }""")
-                    ))
+                    )
+            )
     })
     ResponseEntity<SseEmitter> subscribe(
             @Parameter(hidden = true) CustomOAuth2User principal,
             String lastEventId
     );
+
+    /* ======================  알림 읽음 처리  ====================== */
 
     @Operation(
             summary     = "알림 읽음 처리",
@@ -99,35 +92,22 @@ public interface NotificationControllerSpecification {
             }
     )
     @ApiResponses({
-            @ApiResponse(responseCode = "200",
-                    description  = "처리 성공"),
-            @ApiResponse(responseCode = "400",
-                    description  = "본인 알림이 아님",
-                    content      = @Content(
-                            mediaType = APPLICATION_JSON_VALUE,
-                            examples  = @ExampleObject(value = """
-                            {
-                                "code": 400,
-                                "status": "BAD_REQUEST",
-                                "message": "해당 알림을 읽을 수 없습니다."
-                            }""")
-                    )),
-            @ApiResponse(responseCode = "404",
-                    description  = "알림을 찾을 수 없음",
-                    content      = @Content(
-                            mediaType = APPLICATION_JSON_VALUE,
-                            examples  = @ExampleObject(value = """
-                            {
-                                "code": 404,
-                                "status": "NOT_FOUND",
-                                "message": "알림을 찾을 수 없습니다."
-                            }""")
-                    ))
+            @ApiResponse(responseCode = "200", description = "처리 성공"),
+            @ApiResponse(
+                    responseCode = "400", description = "본인 알림 아님",
+                    content = @Content(mediaType = APPLICATION_JSON_VALUE)
+            ),
+            @ApiResponse(
+                    responseCode = "404", description = "알림 없음",
+                    content = @Content(mediaType = APPLICATION_JSON_VALUE)
+            )
     })
     ResponseEntity<Void> readNotification(
             @Parameter(hidden = true) CustomOAuth2User principal,
             Long notificationId
     );
+
+    /* ======================  전체 / 최근 알림 조회  ====================== */
 
     @Operation(
             summary     = "전체 알림 목록",
@@ -141,33 +121,17 @@ public interface NotificationControllerSpecification {
                     example     = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
             )
     )
-    @ApiResponses({
-            @ApiResponse(responseCode = "200",
-                    description  = "조회 성공",
-                    content      = @Content(
-                            mediaType = APPLICATION_JSON_VALUE,
-                            examples  = @ExampleObject(value = """
-                            [
-                                {
-                                "id": 12,
-                                "title": "새 댓글이 달렸습니다",
-                                "content": "제목: 안녕하세요",
-                                "commentId": 20,
-                                "boardId": 3,
-                                "type": "COMMENT",
-                                "createdAt": "2025.05.22",
-                                "isRead": false
-                                }
-                            ]""")
-                    ))
-    })
+    @ApiResponse(
+            responseCode = "200", description = "조회 성공",
+            content = @Content(mediaType = APPLICATION_JSON_VALUE)
+    )
     ResponseEntity<List<NotificationResponse>> getNotifications(
             @Parameter(hidden = true) CustomOAuth2User principal
     );
 
     @Operation(
             summary     = "최근 알림 3건",
-            description = "가장 최근 3개의 알림만 조회합니다.",
+            description = "가장 최근 3개의 알림을 조회합니다.",
             security    = @SecurityRequirement(name = "AccessCookie"),
             parameters  = @Parameter(
                     name        = "Access",
@@ -177,61 +141,11 @@ public interface NotificationControllerSpecification {
                     example     = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
             )
     )
-    @ApiResponses(@ApiResponse(
-            responseCode = "200",
-            description  = "조회 성공",
-            content      = @Content(mediaType = APPLICATION_JSON_VALUE)
-    ))
+    @ApiResponse(
+            responseCode = "200", description = "조회 성공",
+            content = @Content(mediaType = APPLICATION_JSON_VALUE)
+    )
     ResponseEntity<List<NotificationResponse>> getRecentNotifications(
             @Parameter(hidden = true) CustomOAuth2User principal
-    );
-
-    @Operation(
-            summary     = "댓글 삭제 → 알림 정리",
-            description = "커뮤니티 서비스에서 댓글을 삭제하면 해당 댓글과 관련된 알림을 모두 삭제합니다.",
-            security    = @SecurityRequirement(name = "InternalServiceKey"),   // 내부 호출이라면 별도 security 로 구분
-            parameters  = @Parameter(
-                    name        = "comment-id",
-                    description = "삭제된 댓글 ID",
-                    in          = ParameterIn.PATH,
-                    required    = true,
-                    example     = "42"
-            )
-    )
-    @ApiResponses({
-            @ApiResponse(responseCode = "204", description = "삭제 완료"),
-            @ApiResponse(responseCode = "500", description = "서버 내부 오류")
-    })
-    ResponseEntity<Void> deleteByCommentId(Long commentId);
-
-
-    @Operation(
-            summary     = "댓글 알림 생성",
-            description = "커뮤니티 서비스에서 댓글이 작성될 때 호출되어 알림을 생성합니다.",
-            security    = @SecurityRequirement(name = "InternalServiceKey")
-    )
-    @ApiResponses({
-            @ApiResponse(responseCode = "201", description = "알림 생성 완료"),
-            @ApiResponse(responseCode = "500", description = "서버 내부 오류",
-                    content = @Content(mediaType = APPLICATION_JSON_VALUE))
-    })
-    ResponseEntity<Void> createCommentNotification(
-            @Parameter(description = "댓글 알림 이벤트", required = true)
-            CommentNotificationEvent event
-    );
-
-    @Operation(
-            summary     = "직관팟 알림 생성",
-            description = "TWP 서비스에서 직관팟 관련 이벤트가 발생했을 때 호출되어 알림을 생성합니다.",
-            security    = @SecurityRequirement(name = "InternalServiceKey")
-    )
-    @ApiResponses({
-            @ApiResponse(responseCode = "201", description = "알림 생성 완료"),
-            @ApiResponse(responseCode = "500", description = "서버 내부 오류",
-                    content = @Content(mediaType = APPLICATION_JSON_VALUE))
-    })
-    ResponseEntity<Void> createPartyNotification(
-            @Parameter(description = "직관팟 알림 이벤트", required = true)
-            PartyNotificationEvent event
     );
 }
