+++ src/main/java/com/playus/userservice/global/jwt/JwtFilter.java
@@ -29,6 +29,12 @@ public class JwtFilter extends OncePerRequestFilter {
     private final JwtUtil jwtUtil;
     private final RedisTemplate<String, String> redisTemplate;
 
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException {
+        // 내부 호출 /user/api/ 은 필터에서 제외
+        return request.getRequestURI().startsWith("/user/api/");
+    }
+
     public JwtFilter(JwtUtil jwtUtil,
         RedisTemplate<String, String> redisTemplate) {
         this.jwtUtil = jwtUtil;
